backend/api_foodgram/requirements.txt
django==2.2.16
djangorestframework==3.12.4
django-filter==2.4.0
PyJWT==2.1.0
djangorestframework-simplejwt==5.2.2
django-filter==2.4.0
gunicorn==20.0.4
pytz==2020.1
psycopg2-binary==2.8.6
asgiref==3.2.10
sqlparse==0.3.1
python-dotenv==0.19.0
djoser
Pillow==8.3.1